% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural.R
\name{are_adjacent}
\alias{are_adjacent}
\title{Basic Properties}
\source{
\url{https://igraph.org/c/doc/igraph-Structural.html#basic-properties}
}
\usage{
are_adjacent(graph, from, to)
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}

\item{from}{An integer vector of vertex IDs.}

\item{to}{An integer vector of vertex IDs.}
}
\value{
A logical, \code{TRUE} if \code{from} and \code{to} are adjacent, \code{FALSE} otherwise.
}
\description{
Basic Properties
}
\examples{
g = graph_tree(5L)
are_adjacent(g, 1L, 2L)
are_adjacent(g, 1L, 4L)
}
\seealso{
\code{\link[=edge]{edge()}}, \code{\link[=neighbors]{neighbors()}}, \code{\link[=incident]{incident()}} for similar operations.
\code{\link[=as_adjlist]{as_adjlist()}} and \code{\link[=degree]{degree()}} can apply to all vertices at once.
}
