% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{as_adjlist}
\alias{as_adjlist}
\alias{as_inclist}
\title{Adjacency lists}
\usage{
as_adjlist(graph, mode = 3L)

as_inclist(graph, mode = 3L)
}
\arguments{
\item{graph}{IGraph object.}

\item{mode}{edge type to count; {1: out, 2: in, 3: all}}
}
\description{
These functions are all-vertices version of \code{neighbors()} and \code{incident()}.
}
\details{
\code{as_adjlist()}: Convert a graph to a list of adjacent vertices.

\code{as_inclist()}: Convert a graph to a list of incident edges.
}
\seealso{
\code{neighbors()} and \code{incident()} for single-vertex version.

\url{https://igraph.org/c/doc/ch07.html#igraph-Adjlists}
}
