% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{as_igraph}
\alias{as_igraph}
\alias{as_igraph.default}
\alias{as_igraph.matrix}
\alias{as_igraph.data.frame}
\title{Create graph from R objects}
\usage{
as_igraph(x)

\method{as_igraph}{default}(x)

\method{as_igraph}{matrix}(x)

\method{as_igraph}{data.frame}(x)
}
\arguments{
\item{x}{A vector, matrix, or data.frame.}
}
\description{
Create graph from R objects
}
\seealso{
\code{\link[=graph_create]{graph_create()}}
}
