% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{neighbors}
\alias{neighbors}
\alias{incident}
\alias{degree}
\title{Basic interface}
\usage{
neighbors(graph, v, mode = 3L)

incident(graph, v, mode = 3L)

degree(graph, vids = numeric(0), mode = 3L, loops = TRUE)
}
\arguments{
\item{graph}{IGraph object.}

\item{v}{Vertices of interest.}

\item{mode}{Edge type to count; {1: OUT, 2: IN, 3: ALL}.}

\item{vids}{Vertices of interest.}

\item{loops}{whether self-loop should be counted.}
}
\description{
Basic interface
}
\details{
\code{neighbors()}: Adjacent vertices to a vertex.

\code{incident()}: Gives the incident edges of a vertex.

\code{degree()}: The degree of some vertices in a graph.
}
\seealso{
\code{\link[=as_adjlist]{as_adjlist()}} and \code{\link[=as_inclist]{as_inclist()}} for all-vertices version.

\url{https://igraph.org/c/doc/igraph-Basic.html}
}
