% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliques.R
\name{cliques}
\alias{cliques}
\alias{is_complete}
\alias{is_clique}
\alias{clique_size_hist}
\alias{largest_cliques}
\alias{maximal_cliques}
\alias{maximal_cliques_count}
\alias{maximal_cliques_hist}
\alias{clique_number}
\title{Cliques}
\source{
\url{https://igraph.org/c/doc/igraph-Cliques.html#cliques}
}
\usage{
is_complete(graph)

is_clique(graph, candidate = integer(0L), directed = is_directed(graph))

cliques(graph, min_size = 0L, max_size = -1L, max_results = -1L)

clique_size_hist(graph, min_size = 0L, max_size = -1L)

largest_cliques(graph)

maximal_cliques(graph, min_size = 0L, max_size = -1L, max_results = -1L)

maximal_cliques_count(graph, min_size = 0L, max_size = -1L)

maximal_cliques_hist(graph, min_size = 0L, max_size = -1L)

clique_number(graph)
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}

\item{candidate}{The vertex set to test for being a clique.}

\item{directed}{A logical value, whether to consider directed paths.
Ignored for undirected graphs.}

\item{min_size, max_size}{An integer specifying the minimum/maximum size of the
sets to be returned. If negative or zero, no lower bound will be used.}

\item{max_results}{At most this many sets will be recorded.
If negative, no limit is applied.}
}
\value{
\code{is_complete()} and \code{is_clique()} return a logical value.
}
\description{
A graph is considered complete if all pairs of different vertices are adjacent.
Cliques are fully connected subgraphs of a graph.
}
