% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural.R
\name{subcomponent}
\alias{subcomponent}
\alias{subcomponents}
\alias{induced_subgraph}
\title{Graph Components}
\source{
\url{https://igraph.org/c/doc/igraph-Structural.html#graph-components}
}
\usage{
subcomponent(graph, vid, mode = 1L)

subcomponents(graph, vids, mode = 1L)

induced_subgraph(graph, vids, impl = 0L)
}
\arguments{
\item{graph}{igraph_ptr object.}

\item{vid}{Vertices of interest.}

\item{mode}{Edge type to count; \{1: OUT, 2: IN, 3: ALL\}.}

\item{vids}{Vertices of interest.}

\item{impl}{how to construct a new graph:
\{0: AUTO, 1: COPY_AND_DELETE, 2: CREATE_FROM_SCRATCH\}}
}
\description{
Graph Components
}
