% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural.R
\name{are_connected}
\alias{are_connected}
\title{Basic Properties}
\source{
\url{https://igraph.org/c/doc/igraph-Structural.html#idm470927297936}
}
\usage{
are_connected(graph, from, to)
}
\arguments{
\item{graph}{IGraph object.}

\item{from}{Vertices of interest.}

\item{to}{Vertices of interest.}
}
\description{
Basic Properties
}
