% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{graph_from_edgelist}
\alias{graph_from_edgelist}
\alias{graph_from_symbolic_edgelist}
\title{Create new graph}
\usage{
graph_from_edgelist(el, directed = TRUE)

graph_from_symbolic_edgelist(el, directed = TRUE)
}
\arguments{
\item{el}{two-column matrix}

\item{directed}{logical}
}
\description{
Create new graph
}
