% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion-impl.R
\name{graph_from_data_frame}
\alias{graph_from_data_frame}
\alias{graph_from_symbolic_edgelist}
\alias{graph_from_edgelist}
\title{Convert edge lists to igraph objects.}
\usage{
graph_from_data_frame(edgelist, directed = TRUE)

graph_from_symbolic_edgelist(edgelist, directed = TRUE)

graph_from_edgelist(edgelist, directed = TRUE)
}
\arguments{
\item{edgelist}{An edge list in two-column matrix or data.frame format.
\code{graph_from_edgelist()} only accepts integer values,
interpreting them directly as vertex IDs starting from 1.
Other functions support non-numeric values, treating them as vertex names.
A data.frame may contain additional columns for edge attributes.}

\item{directed}{A logical.}
}
\value{
An \link{igraph_ptr} object.
}
\description{
An edge list is a two-column matrix or data.frame where the first column
contains the source vertices and the second column contains the target
vertices.
}
\examples{
edges = seq_len(6L) + 3L
edgelist = matrix(edges, ncol = 2L, byrow = TRUE)
g = graph_from_edgelist(edgelist) |> print()
stopifnot(all.equal(as.matrix(g), as.matrix(graph_create(edges))))

symbolic = matrix(letters[edges], ncol = 2L)
graph_from_symbolic_edgelist(symbolic)
}
\seealso{
\code{\link[=as.data.frame.igraph_ptr]{as.data.frame.igraph_ptr()}} and \code{\link[=as.matrix.igraph_ptr]{as.matrix.igraph_ptr()}}
do the opposite conversion.
}
