% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as_igraph}
\alias{as_igraph}
\alias{as_igraph.default}
\alias{as_igraph.matrix}
\alias{as_igraph.data.frame}
\alias{as.data.frame.igraph_ptr}
\alias{as.matrix.igraph_ptr}
\title{Convert igraph objects to and from other formats.}
\usage{
as_igraph(x, ...)

\method{as_igraph}{default}(x, ...)

\method{as_igraph}{matrix}(x, ...)

\method{as_igraph}{data.frame}(x, ...)

\method{as.data.frame}{igraph_ptr}(x, ...)

\method{as.matrix}{igraph_ptr}(x, ...)
}
\arguments{
\item{x}{A vector, matrix, data.frame, or graph object to be converted.}

\item{...}{Additional arguments passed to the conversion function.}
}
\value{
\code{as_igraph()} returns an igraph object.

\code{as.data.frame()} returns an edge list in data.frame format,
which may include additional columns for edge attributes.

\code{as.matrix()} returns an edge list in matrix format.
}
\description{
Round-trip conversion is not guaranteed currently.
Conversion to data.frame preserves edge attributes and vertex names,
but the other vertex attributes are lost, and internal vertex IDs may change.
}
\examples{
edges = seq_len(6L)
g1 = as_igraph(edges)
edgelist = as.matrix(g1) |> print()

g2 = as_igraph(edgelist)
g3 = as_igraph(as.data.frame(g2)) |> print()
}
\seealso{
\code{\link[=graph_from_symbolic_edgelist]{graph_from_symbolic_edgelist()}}, \code{\link[=graph_from_edgelist]{graph_from_edgelist()}},
\code{\link[=graph_from_data_frame]{graph_from_data_frame()}}, \code{\link[=graph_create]{graph_create()}} for underlying functions.
}
