% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{is_dag}
\alias{is_dag}
\alias{is_acyclic}
\title{Acyclic graphs and feedback sets}
\source{
\url{https://igraph.org/c/doc/igraph-Cycles.html#acyclic-graphs-feedback-sets}
}
\usage{
is_dag(graph)

is_acyclic(graph)
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}
}
\value{
A logical value.
}
\description{
A directed acyclic graph (DAG) is a directed graph with no directed cycles.
For directed graphs, only directed cycles are considered.
}
\examples{
dag = graph_tree(5L)
tree = graph_tree(5L, mode = 2L)
ring = graph_ring(5L)

is_dag(dag)
is_dag(tree)
is_dag(ring)

is_acyclic(dag)
is_acyclic(tree)
is_acyclic(ring)
}
