% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycles.R
\name{is_eulerian}
\alias{is_eulerian}
\title{Eulerian cycles and paths}
\source{
\url{https://igraph.org/c/doc/igraph-Cycles.html#eulerian-cycles}
}
\usage{
is_eulerian(graph)
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}
}
\value{
A logical value with two elements: "has a path" and "has a cycle".
}
\description{
An Eulerian path traverses each edge of the graph precisely once.
A closed Eulerian path is referred to as an Eulerian cycle.
}
\examples{
is_eulerian(graph_tree(5L))

is_eulerian(graph_square_lattice(5L))

is_eulerian(graph_ring(5L))
}
