% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{degree}
\alias{degree}
\title{The degree of some vertices in a graph}
\usage{
degree(graph, vids = numeric(0), mode = 3L, loops = TRUE)
}
\arguments{
\item{graph}{IGraph object.}

\item{vids}{vertices of which the degree will be calculated.}

\item{mode}{edge type to count; {1: out, 2: in, 3: all}}

\item{loops}{whether self-loop should be counted.}
}
\description{
This function calculates the in-, out- or total degree of the specified vertices.
}
