% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-distances.R
\name{distances}
\alias{distances}
\alias{average_path_length}
\alias{path_length_hist}
\title{Distance-related Functions}
\source{
\url{https://igraph.org/c/doc/igraph-Structural.html#shortest-path-related-functions}
}
\usage{
distances(
  graph,
  from = integer(0L),
  to = from,
  weights = numeric(0L),
  mode = 3L,
  algorithm = c("dijkstra", "bellman-ford", "johnson")
)

average_path_length(
  graph,
  weights = numeric(0L),
  directed = is_directed(graph),
  unconn = TRUE
)

path_length_hist(graph, directed = is_directed(graph))
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}

\item{from}{An integer vector of vertex IDs.}

\item{to}{An integer vector of vertex IDs.}

\item{weights}{A numeric vector of edge weights; \code{TRUE} to use \code{Eattr(graph, "weight")}.}

\item{mode}{An integer value of edge type to count; \{1: OUT, 2: IN, 3: ALL\}.}

\item{algorithm}{A character string.}

\item{directed}{A logical value, whether to consider directed paths.
Ignored for undirected graphs.}

\item{unconn}{A logical value.}
}
\value{
\code{distances()} returns a matrix of distances between the vertices.

\code{average_path_length()} returns a numeric value.

\code{path_length_hist()} returns an integer vector.
}
\description{
\code{average_path_length()} and \code{path_length_hist()} can be computed with the
result of \code{distances()}, but should be faster as they are written in C.
}
\examples{
g = graph_tree(5L)
distances(g, mode = 1L)
average_path_length(g)
path_length_hist(g)
}
