% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-efficiency.R
\name{global_efficiency}
\alias{global_efficiency}
\alias{local_efficiency}
\alias{average_local_efficiency}
\title{Efficiency measures}
\source{
\url{https://igraph.org/c/doc/igraph-Structural.html#efficiency-measures}
}
\usage{
global_efficiency(graph, weights = numeric(0L), directed = is_directed(graph))

local_efficiency(
  graph,
  weights = numeric(0L),
  vids = integer(0L),
  directed = is_directed(graph),
  mode = 3L
)

average_local_efficiency(
  graph,
  weights = numeric(0L),
  directed = is_directed(graph),
  mode = 3L
)
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}

\item{weights}{A numeric vector of edge weights; \code{TRUE} to use \code{Eattr(graph, "weight")}.}

\item{directed}{A logical value, whether to consider directed paths.
Ignored for undirected graphs.}

\item{vids}{An integer vector of vertex IDs.}

\item{mode}{An integer value of edge type to count; \{1: OUT, 2: IN, 3: ALL\}.}
}
\value{
A numeric vector.
}
\description{
The global efficiency of a network is defined as the average of inverse distances.
The local efficiency of a network around a vertex is the global efficiency of the
subgraph after removing the vertex.
}
\examples{
g = graph_tree(7L, mode = 2L)
global_efficiency(g)

local_efficiency(g)

average_local_efficiency(g)
}
