% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{igraph_from}
\alias{igraph_from}
\alias{igraph_to}
\title{Quick access to edge list columns}
\source{
\url{https://github.com/igraph/igraph/blob/master/include/igraph_datatype.h}
}
\usage{
igraph_from(graph)

igraph_to(graph)
}
\arguments{
\item{graph}{igraph_ptr object.}
}
\value{
\code{igraph_from()} returns the head vertices of all the edges.

\code{igraph_to()} returns the tail vertices of all the edges.
}
\description{
If only one of the two columns in the edge list is needed,
these functions provide a more efficient way to access them.
}
\examples{
g = graph_create(seq_len(6L))
igraph_from(g)
igraph_to(g)
}
\seealso{
\code{\link[=as.matrix.igraph_ptr]{as.matrix.igraph_ptr()}} and \code{\link[=as.data.frame.igraph_ptr]{as.data.frame.igraph_ptr()}} to
get entire edge lists.
}
