% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{watts_strogatz_game}
\alias{watts_strogatz_game}
\alias{rewire_edges}
\title{Edge rewiring models}
\source{
\url{https://igraph.org/c/html/latest/igraph-Games.html#edge-rewiring-games}
}
\usage{
watts_strogatz_game(dim, size, nei, p, edge_type_sw = 0L)

rewire_edges(graph, p, edge_type_sw = 0L)
}
\arguments{
\item{dim}{The dimension of the lattice.}

\item{size}{The size of the lattice along each dimension.}

\item{nei}{The size of the neighborhood for each vertex.}

\item{p}{The probability of edge creation or rewiring.}

\item{edge_type_sw}{The types of edges to allow in the graph:
\{0: simple, 1: loops, 6: multiple, 7: loops and multiple\}.}

\item{graph}{An \code{\link{igraph_ptr}} object.}
}
\value{
An \code{\link{igraph_ptr}} object.
}
\description{
\code{rewire_edges()} creates a copy of the input graph,
and rewires edges randomly with probability \code{p} for each edge.
}
\examples{
regular = watts_strogatz_game(1L, 20L, 2L, 0)
random = watts_strogatz_game(1L, 20L, 2L, 1)
small_world = watts_strogatz_game(1L, 20L, 2L, 0.1)
small_world2 = rewire_edges(regular, 0.1)
}
\seealso{
\code{\link[=rewire]{rewire()}} for the degree-preserving edge rewiring model.
}
