% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/games.R
\name{erdos_renyi_game_gnm}
\alias{erdos_renyi_game_gnm}
\alias{erdos_renyi_game_gnp}
\title{The Erdős-Rényi and related models}
\source{
\url{https://igraph.org/c/html/latest/igraph-Games.html#erdos-renyi-games}
}
\usage{
erdos_renyi_game_gnm(n, m, directed = FALSE, edge_type_sw = 0L)

erdos_renyi_game_gnp(n, p, directed = FALSE, edge_type_sw = 0L)
}
\arguments{
\item{n}{The number of vertices.}

\item{m}{The number of edges.}

\item{directed}{A logical value, whether to consider directed paths.
Ignored for undirected graphs.}

\item{edge_type_sw}{The types of edges to allow in the graph:
\{0: simple, 1: loops, 6: multiple, 7: loops and multiple\}.}

\item{p}{The probability of edge creation or rewiring.}
}
\value{
An \code{\link{igraph_ptr}} object.
}
\description{
Classical random graph models with a fixed number of vertices.
}
\examples{
gnm = erdos_renyi_game_gnm(5L, 5L)
gnp = erdos_renyi_game_gnp(5L, 0.5)
}
