% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{graph_create}
\alias{graph_create}
\title{Create a graph with the specified edges}
\usage{
graph_create(edges, n = 0L, directed = TRUE)
}
\arguments{
\item{edges}{An integer vector of vertex IDs; the first two elements are the first edge, etc.}

\item{n}{An integer value, the number of vertices.}

\item{directed}{A logical value, whether to consider directed paths.
Ignored for undirected graphs.}
}
\value{
An \code{\link{igraph_ptr}} object.
}
\description{
Create a graph with the specified edges
}
\examples{
graph_create(seq_len(6L))
}
\seealso{
\code{\link[=as_igraph]{as_igraph()}}, \code{\link[=graph_from_data_frame]{graph_from_data_frame()}}

Other deterministic generators: 
\code{\link{graph_empty}()},
\code{\link{graph_famous}()},
\code{\link{graph_full}()},
\code{\link{graph_hypercube}()},
\code{\link{graph_kary_tree}()},
\code{\link{graph_lcf}()},
\code{\link{graph_square_lattice}()},
\code{\link{graph_star}()}
}
\concept{deterministic generators}
