% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{graph_famous}
\alias{graph_famous}
\title{Create a famous graph by simply providing its name}
\source{
\url{https://igraph.org/c/doc/igraph-Generators.html#igraph_famous}
}
\usage{
graph_famous(name)
}
\arguments{
\item{name}{A character string, the name of the graph.}
}
\value{
An \code{\link{igraph_ptr}} object.
}
\description{
See the C library documentation for the list of available graphs.
}
\examples{
graph_famous("Diamond")
}
\seealso{
Other deterministic generators: 
\code{\link{graph_create}()},
\code{\link{graph_empty}()},
\code{\link{graph_full}()},
\code{\link{graph_kary_tree}()},
\code{\link{graph_square_lattice}()},
\code{\link{graph_star}()}
}
\concept{deterministic generators}
