% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{graph_full}
\alias{graph_full}
\title{Create a full graph (complete graph)}
\usage{
graph_full(n, directed = FALSE, self_loops = FALSE)
}
\arguments{
\item{n}{An integer value, the number of vertices.}

\item{directed}{A logical value, whether to consider directed paths.
Ignored for undirected graphs.}

\item{self_loops}{A logical value, whether to include self-loops.}
}
\value{
An \code{\link{igraph_ptr}} object.
}
\description{
In a full graph, every vertex is connected to every other vertex.
}
\examples{
graph_full(4L)
graph_full(4L, directed = TRUE)
}
\seealso{
Other deterministic generators: 
\code{\link{graph_create}()},
\code{\link{graph_empty}()},
\code{\link{graph_famous}()},
\code{\link{graph_kary_tree}()},
\code{\link{graph_square_lattice}()},
\code{\link{graph_star}()}
}
\concept{deterministic generators}
