% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{graph_hypercube}
\alias{graph_hypercube}
\title{Create an n-dimensional hypercube graph}
\source{
\url{https://igraph.org/c/doc/igraph-Generators.html#igraph_hypercube}
}
\usage{
graph_hypercube(n, directed = FALSE)
}
\arguments{
\item{n}{The dimension of the hypercube graph.}

\item{directed}{A logical value, whether to consider directed paths.
Ignored for undirected graphs.}
}
\value{
An \code{\link{igraph_ptr}} object.
}
\description{
The hypercube graph \code{Q_n} has \code{2^n} vertices and \verb{2^(n-1) n} edges.
}
\examples{
graph_hypercube(3L)
}
\seealso{
Other deterministic generators: 
\code{\link{graph_create}()},
\code{\link{graph_empty}()},
\code{\link{graph_famous}()},
\code{\link{graph_full}()},
\code{\link{graph_kary_tree}()},
\code{\link{graph_lcf}()},
\code{\link{graph_square_lattice}()},
\code{\link{graph_star}()}
}
\concept{deterministic generators}
