% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{graph_lcf}
\alias{graph_lcf}
\alias{graph_circulant}
\title{Creates a graph from LCF notation}
\source{
\url{https://igraph.org/c/html/latest/igraph-Generators.html#igraph_lcf}
}
\usage{
graph_lcf(n, shifts, repeats = n)

graph_circulant(n, shifts, directed = FALSE)
}
\arguments{
\item{n}{An integer value, the number of vertices.}

\item{shifts}{An integer vector giving the shifts.
For \code{graph_lcf()}, it gives additional edges to a cycle backbone,
while for \code{graph_circulant()}, it gives all the edges.}

\item{repeats}{The number of repeats for the shifts.}

\item{directed}{A logical value, whether to consider directed paths.
Ignored for undirected graphs.}
}
\value{
An \code{\link{igraph_ptr}} object.
}
\description{
LCF notation (named after Lederberg, Coxeter, and Frucht) is a concise notation
for 3-regular Hamiltonian graphs.

\code{graph_circulant(n, shifts)} is roughly equivalent to \code{graph_lcf(n, c(1L, shifts), n)}.
}
\examples{
graph_lcf(5L, 2L, 5L)

graph_circulant(5L, 2L, directed = TRUE)
}
\seealso{
Other deterministic generators: 
\code{\link{graph_create}()},
\code{\link{graph_empty}()},
\code{\link{graph_famous}()},
\code{\link{graph_full}()},
\code{\link{graph_hypercube}()},
\code{\link{graph_kary_tree}()},
\code{\link{graph_square_lattice}()},
\code{\link{graph_star}()}
}
\concept{deterministic generators}
