% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{graph_star}
\alias{graph_star}
\title{Create a star graph}
\usage{
graph_star(n, mode = 0L, center = 1L)
}
\arguments{
\item{n}{An integer value, the number of vertices.}

\item{mode}{An integer value, \{0: OUT, 1: IN, 2: UNDIRECTED, 3: MUTUAL\}}

\item{center}{Id of the vertex which will be the center of the graph.}
}
\value{
An \code{\link{igraph_ptr}} object.
}
\description{
In a star graph, every vertex connects only to the center.
}
\examples{
graph_star(4L, 0L)
graph_star(4L, 1L)
graph_star(4L, 2L)
graph_star(4L, 3L)
}
\seealso{
Other deterministic generators: 
\code{\link{graph_create}()},
\code{\link{graph_empty}()},
\code{\link{graph_famous}()},
\code{\link{graph_full}()},
\code{\link{graph_kary_tree}()},
\code{\link{graph_square_lattice}()}
}
\concept{deterministic generators}
