% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{graph_kary_tree}
\alias{graph_kary_tree}
\alias{graph_tree}
\title{Create a tree graph}
\usage{
graph_kary_tree(n, children = 2L, mode = 0L)

# graph_tree = graph_kary_tree
}
\arguments{
\item{n}{An integer value, the number of vertices.}

\item{children}{Integer, the number of children of a vertex in the tree.}

\item{mode}{An integer value,
\code{igraph_tree_mode_t} \{0: OUT, 1: IN, 2: UNDIRECTED\}}
}
\value{
An \code{\link{igraph_ptr}} object.
}
\description{
A k-ary tree is a tree where each vertex has up to \emph{k} children.
The symmetry of the graph depends on \code{n}.
}
\examples{
graph_kary_tree(5L)
}
\seealso{
Other deterministic generators: 
\code{\link{graph_create}()},
\code{\link{graph_empty}()},
\code{\link{graph_famous}()},
\code{\link{graph_full}()},
\code{\link{graph_square_lattice}()},
\code{\link{graph_star}()}
}
\concept{deterministic generators}
