% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraphlite-package.R
\docType{class}
\name{IGraph}
\alias{IGraph}
\title{IGraph class: thin wrapper of \code{igraph_t}.}
\format{
Reference Class (RC) exposed by Rcpp Modules.
}
\source{
\url{https://github.com/igraph/igraph/blob/master/include/igraph_datatype.h}
}
\description{
IGraph class: thin wrapper of \code{igraph_t}.
}
\seealso{
\code{\link[=graph_from_edgelist]{graph_from_edgelist()}}, \link{graph_from_data_frame}, \code{\link[=graph_tree]{graph_tree()}}
for object instantiation.
}
\keyword{data}
