% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraphlite-package.R
\docType{package}
\name{igraphlite-package}
\alias{igraphlite}
\alias{igraphlite-package}
\title{igraphlite: R interface to C igraph}
\description{
R interface to C igraph.
}
\author{
\strong{Maintainer}: Watal M. Iwasaki \email{heavy.watal@gmail.com}

}
\keyword{internal}
