% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliques.R
\name{independent}
\alias{independent}
\alias{is_independent_vertex_set}
\alias{independent_vertex_sets}
\alias{largest_independent_vertex_sets}
\alias{maximal_independent_vertex_sets}
\alias{independence_number}
\title{Independent vertex sets}
\source{
\url{https://igraph.org/c/doc/igraph-Cliques.html#independent-vertex-sets}
}
\usage{
is_independent_vertex_set(graph, candidate = integer(0L))

independent_vertex_sets(
  graph,
  min_size = 0L,
  max_size = -1L,
  max_results = -1L
)

largest_independent_vertex_sets(graph)

maximal_independent_vertex_sets(
  graph,
  min_size = 0L,
  max_size = -1L,
  max_results = -1L
)

independence_number(graph)
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}

\item{candidate}{The vertex set to test for being a clique.}

\item{min_size, max_size}{An integer specifying the minimum/maximum size of the
sets to be returned. If negative or zero, no lower bound will be used.}

\item{max_results}{At most this many sets will be recorded.
If negative, no limit is applied.}
}
\description{
A vertex set is considered independent if there are no edges between them.
}
