% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isomorphism.R
\name{isomorphic}
\alias{isomorphic}
\alias{subisomorphic}
\title{The simple interface for graph isomorphism}
\source{
\url{https://igraph.org/c/html/latest/igraph-Isomorphism.html#isomorphism-simple-interface}
}
\usage{
isomorphic(graph1, graph2)

subisomorphic(graph1, graph2)
}
\arguments{
\item{graph1, graph2}{Two \code{\link{igraph_ptr}} objects to compare.}
}
\value{
A logical value.

\code{subisomorphic()} returns \code{TRUE} if \code{graph2} is isomorphic
to a subgraph of \code{graph1}.
}
\description{
In simple terms, two graphs are isomorphic if they become indistinguishable
from each other once their vertex labels are removed.
}
\examples{
g1 = graph_create(c(1L, 2L, 1L, 3L), directed = FALSE)
g2 = graph_create(c(3L, 2L, 3L, 1L), directed = FALSE)
isomorphic(g1, g2)
subisomorphic(g1, g2)

g3 = graph_create(c(1L, 2L, 1L, 3L, 3L, 4L), directed = FALSE)
subisomorphic(g3, g1)
subisomorphic(g1, g3)
}
\seealso{
\code{\link[=is_same_graph]{is_same_graph()}}.
}
