% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_reingold_tilford}
\alias{layout_reingold_tilford}
\alias{layout_reingold_tilford_circular}
\title{Layouts for trees and acyclic graphs}
\source{
\url{https://igraph.org/c/doc/igraph-Layout.html#layouts-for-trees-and-acyclic-graphs}
}
\usage{
layout_reingold_tilford(graph, mode = 1L, roots = integer(0))

layout_reingold_tilford_circular(graph, mode = 1L, roots = integer(0))
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}

\item{mode}{An integer value of edge type to count; \{1: OUT, 2: IN, 3: ALL\}.}

\item{roots}{An integer vector of vertex IDs, the roots of the tree.}
}
\value{
An \code{\link{igraph_ptr}} object with vertex attributes \code{x} and \code{y}.
}
\description{
Layout functions add \code{x} and \code{y} vertex attributes to the graph object.
The object is modified in place, and its shallow copy is returned.
}
\examples{
g = graph_tree(5L)

layout_reingold_tilford(g) |> Vattr()

layout_reingold_tilford_circular(g) |> Vattr()
}
\seealso{
Other layouts: 
\code{\link{layout_nicely}()},
\code{\link{layout_random}()}
}
\concept{layouts}
