% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_nicely}
\alias{layout_nicely}
\title{Choose a layout algorithm automatically}
\source{
\url{https://r.igraph.org/reference/layout_nicely.html}
}
\usage{
layout_nicely(graph, ...)
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}

\item{...}{Additional arguments passed to the layout function.}
}
\value{
An \code{\link{igraph_ptr}} object with vertex attributes \code{x} and \code{y}.
}
\description{
Layout functions add \code{x} and \code{y} vertex attributes to the graph object.
The object is modified in place, and its shallow copy is returned.
}
\examples{
layout_nicely(graph_tree(5L)) |> Vattr()
}
\seealso{
Other layouts: 
\code{\link{layout_random}()},
\code{\link{layout_reingold_tilford}()}
}
\concept{layouts}
