% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-non-simple.R
\name{is_simple}
\alias{is_simple}
\alias{has_loop}
\alias{count_loops}
\alias{has_multiple}
\alias{count_multiple}
\title{Non-simple graphs: multiple and loop edges}
\source{
\url{https://igraph.org/c/doc/igraph-Structural.html#non-simple-graphs-multiple-and-loop-edges}
}
\usage{
is_simple(graph, directed = is_directed(graph))

has_loop(graph)

count_loops(graph)

has_multiple(graph)

count_multiple(graph, eids = integer(0L))
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}

\item{directed}{A logical value, whether to consider directed paths.
Ignored for undirected graphs.}

\item{eids}{An integer vector of edge IDs.}
}
\value{
Logical or integer values.
}
\description{
Multiple edges are those that connect the same pair of vertices, where the direction
is also taken into account in directed graphs.
Loop here means auto-loop/self-loop, i.e., an edge that connects a vertex to itself.
}
\examples{
g = graph_tree(5L)
g = add_edges(g, c(1L, 1L)) # add a loop
g = add_edges(g, c(1L, 2L)) # add a multiple edge
g = add_edges(g, c(2L, 1L)) # not a multiple edge in directed graph

is_simple(g)
has_loop(g)
count_loops(g)
has_multiple(g)
count_multiple(g)
}
