% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-paths.R
\name{paths}
\alias{paths}
\alias{get_shortest_paths}
\alias{get_all_shortest_paths}
\alias{get_all_simple_paths}
\title{Shortest Path Related Functions}
\source{
\url{https://igraph.org/c/doc/igraph-Structural.html#shortest-path-related-functions}
}
\usage{
get_shortest_paths(
  graph,
  from,
  to = integer(0L),
  weights = numeric(0L),
  mode = 3L
)

get_all_shortest_paths(
  graph,
  from,
  to = integer(0L),
  weights = numeric(0L),
  mode = 3L
)

get_all_simple_paths(graph, from, to = integer(0L), cutoff = -1L, mode = 3L)
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}

\item{from}{An integer vector of vertex IDs.}

\item{to}{An integer vector of vertex IDs.}

\item{weights}{A numeric vector of edge weights; \code{TRUE} to use \code{Eattr(graph, "weight")}.}

\item{mode}{An integer value of edge type to count; \{1: OUT, 2: IN, 3: ALL\}.}

\item{cutoff}{integer}
}
\value{
A list of integer IDs of vertices in the paths.
}
\description{
\code{get_shortest_paths()} returns a single shortest path per vertex pair,
while the variant with \verb{_all_} returns all found paths.
The variant with \verb{_simple_} removes paths with cycles.
The results should be the same for simple graphs like trees, but the order of
the paths may differ.
}
\examples{
g = graph_tree(5L)
vsource = Vsource(g)
vsink = Vsink(g)
get_shortest_paths(g, vsource, vsink, mode = 1L)
get_all_shortest_paths(g, vsource, vsink, mode = 1L)
get_all_simple_paths(g, vsource, vsink, mode = 1L)
}
