% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{as_igraph.phylo}
\alias{as_igraph.phylo}
\alias{as_phylo}
\title{Conversion between phylo and igraph}
\source{
\url{https://cran.r-project.org/package=ape}
}
\usage{
\method{as_igraph}{phylo}(x, ...)

as_phylo(x)
}
\arguments{
\item{x}{An object to convert.}

\item{...}{Additional arguments passed to \code{\link[=graph_from_edgelist]{graph_from_edgelist()}}.}
}
\description{
Conversion between phylo and igraph
}
