% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-method.R
\name{augment.igraph_ptr}
\alias{augment.igraph_ptr}
\alias{plot.igraph_ptr}
\title{Methods for quick visualization}
\usage{
\method{augment}{igraph_ptr}(x, layout = NULL, ...)

\method{plot}{igraph_ptr}(x, ..., lwd = 0.5, cex = 5, col = "#cccccc", pch = 16)
}
\arguments{
\item{x}{igraph_ptr object.}

\item{layout}{A function or resulting data.frame.
If not provided, \code{\link[=layout_nicely]{layout_nicely()}} is applied.}

\item{...}{passed to \code{\link[=augment]{augment()}} or layout function.}

\item{lwd}{passed to \code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}}.}

\item{cex, col, pch}{passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} and \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}.}
}
\description{
Methods for quick visualization
}
