% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign.R
\name{read_graph}
\alias{read_graph}
\alias{read_graph_edgelist}
\alias{read_graph_ncol}
\alias{read_graph_lgl}
\alias{read_graph_graphml}
\alias{read_graph_gml}
\alias{read_graph_pajek}
\alias{read_graph_dl}
\title{Read graphs from files}
\source{
\url{https://igraph.org/c/doc/igraph-Foreign.html}
}
\usage{
read_graph(
  file,
  format = c("edgelist", "ncol", "lgl", "graphml", "gml", "pajek", "dl"),
  ...
)

read_graph_edgelist(file, directed = FALSE)

read_graph_ncol(file, names = FALSE, directed = FALSE)

read_graph_lgl(file, names = FALSE, directed = FALSE)

read_graph_graphml(file, index = 0L)

read_graph_gml(file)

read_graph_pajek(file)

read_graph_dl(file, directed = FALSE)
}
\arguments{
\item{file}{A string specifying the file path.}

\item{format}{Estimated from the file extension if not provided.}

\item{...}{Additional arguments passed to specific reading functions.}

\item{directed}{A logical value, whether to consider directed paths.
Ignored for undirected graphs.}

\item{names}{A logical value indicating whether to read vertex names.}

\item{index}{An integer specifying which graph to read (starting from 0).}
}
\value{
A \code{\link{igraph_ptr}} object.
}
\description{
Currently attributes are not supported.
}
\seealso{
\code{\link[=write_graph]{write_graph()}}.
}
