% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-method.R
\name{as.data.frame.igraph_ptr}
\alias{as.data.frame.igraph_ptr}
\alias{as.matrix.igraph_ptr}
\alias{print.igraph_ptr}
\title{S3 methods to get and view internal data.}
\usage{
\method{as.data.frame}{igraph_ptr}(x, ...)

\method{as.matrix}{igraph_ptr}(x, ...)

\method{print}{igraph_ptr}(x, ...)
}
\arguments{
\item{x}{igraph_ptr object.}

\item{...}{passed to methods.}
}
\description{
S3 methods to get and view internal data.
}
\seealso{
\code{\link[=graph_from_data_frame]{graph_from_data_frame()}}, \code{\link[=graph_from_symbolic_edgelist]{graph_from_symbolic_edgelist()}}
}
