% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural.R
\name{shortest_paths}
\alias{shortest_paths}
\alias{get_shortest_paths}
\alias{average_path_length}
\alias{path_length_hist}
\title{Shortest Path Related Functions}
\source{
\url{https://igraph.org/c/doc/igraph-Structural.html#idm470927281168}
}
\usage{
shortest_paths(graph, from = numeric(0L), to = from,
  weights = numeric(0L), mode = 3L, algorithm = c("auto",
  "unweighted", "dijkstra", "bellman-ford", "johnson"))

get_shortest_paths(graph, from, to = numeric(0L),
  weights = numeric(0L), mode = 3L)

average_path_length(graph, directed = FALSE)

path_length_hist(graph, directed = FALSE)
}
\arguments{
\item{graph}{IGraph object.}

\item{from}{Vertices of interest.}

\item{to}{Vertices of interest.}

\item{weights}{Numeric edge weights; \code{TRUE} to use \code{graph$Eattr$weight}.}

\item{mode}{Edge type to count; {1: OUT, 2: IN, 3: ALL}.}

\item{algorithm}{character}

\item{directed}{Boolean, whether to consider directed paths. Ignored for undirected graphs.}
}
\description{
Shortest Path Related Functions
}
