% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink.R
\name{is_sink}
\alias{is_sink}
\alias{is_source}
\alias{Vsink}
\alias{Vsource}
\title{Shortcuts to get sink/source vertices.}
\usage{
is_sink(graph)

is_source(graph)

Vsink(graph)

Vsource(graph)
}
\arguments{
\item{graph}{igraph_ptr object.}
}
\description{
Shortcuts to get sink/source vertices.
}
