% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{induced_subgraph}
\alias{induced_subgraph}
\alias{induced_subgraph_edges}
\alias{subgraph_from_edges}
\title{Miscellaneous operators}
\source{
\url{https://igraph.org/c/doc/igraph-Operators.html#miscellaneous-operators}
}
\usage{
induced_subgraph(graph, vids, impl = 0L)

induced_subgraph_edges(graph, vids)

subgraph_from_edges(graph, eids, delete_vertices = TRUE)
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}

\item{vids}{An integer vector of vertex IDs.}

\item{impl}{An integer value, how to construct a new graph:
\{0: AUTO, 1: COPY_AND_DELETE, 2: CREATE_FROM_SCRATCH\}}

\item{eids}{An integer vector of edge IDs.}

\item{delete_vertices}{whether to delete the unused vertices.}
}
\value{
An \code{\link{igraph_ptr}} object.
}
\description{
\code{induced_subgraph()} can be achieved in two steps with
\code{induced_subgraph_edges()} and \code{subgraph_from_edges()},
but it is much slower than expected.
}
\examples{
g = graph_tree(5L)
induced_subgraph(g, seq_len(3L))

edges = induced_subgraph_edges(g, seq_len(3L)) |> print()

subgraph_from_edges(g, edges)
}
\seealso{
\code{\link[=delete_vertices]{delete_vertices()}}
}
