% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_tree.R
\name{as_igraph.tbl_tree}
\alias{as_igraph.tbl_tree}
\alias{as_tbl_tree}
\title{Conversion between tbl_tree and igraph}
\source{
\url{https://cran.r-project.org/package=tidytree}
}
\usage{
\method{as_igraph}{tbl_tree}(x, ...)

as_tbl_tree(x)
}
\arguments{
\item{x}{An object to convert.}

\item{...}{Additional arguments passed to \code{\link[=graph_from_edgelist]{graph_from_edgelist()}}.}
}
\description{
Conversion between tbl_tree and igraph
}
