% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{Vnames}
\alias{Vnames}
\alias{as_vids}
\alias{as_vnames}
\title{Convesion between vertex IDs and names}
\usage{
Vnames(graph)

as_vids(graph, vnames)

as_vnames(graph, vids = graph$V)
}
\arguments{
\item{graph}{IGraph object.}

\item{vnames}{Vertex names stored in \code{Vattr[["name"]]}.}

\item{vids}{Vertices of interest.}
}
\description{
Convesion between vertex IDs and names
}
