% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign.R
\name{write_graph}
\alias{write_graph}
\alias{write_graph_edgelist}
\alias{write_graph_ncol}
\alias{write_graph_lgl}
\alias{write_graph_graphml}
\alias{write_graph_gml}
\alias{write_graph_pajek}
\alias{write_graph_dot}
\alias{write_graph_leda}
\title{Write graphs to files}
\source{
\url{https://igraph.org/c/doc/igraph-Foreign.html}
}
\usage{
write_graph(
  graph,
  file,
  format = c("edgelist", "ncol", "lgl", "graphml", "gml", "pajek", "dot", "leda"),
  ...
)

write_graph_edgelist(graph, file)

write_graph_ncol(graph, file)

write_graph_lgl(graph, file, isolates = TRUE)

write_graph_graphml(graph, file, prefixattr = FALSE)

write_graph_gml(graph, file, creator = "igraphlite")

write_graph_pajek(graph, file)

write_graph_dot(graph, file)

write_graph_leda(graph, file)
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}

\item{file}{A string specifying the file path.}

\item{format}{Estimated from the file extension if not provided.}

\item{...}{Additional arguments passed to specific writing functions.}

\item{isolates}{A logical value indicating whether to include isolated vertices.}

\item{prefixattr}{A logical value indicating whether to fix non-unique attribute names.}

\item{creator}{A string specifying the creator of the GML file.}
}
\value{
The input graph is returned invisibly.
}
\description{
Currently attributes are not supported.
}
\seealso{
\code{\link[=read_graph]{read_graph()}}.
}
