% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjlist.R
\name{as_adjlist}
\alias{as_adjlist}
\alias{as_inclist}
\title{Adjacency lists}
\source{
\url{https://igraph.org/c/doc/igraph-Data-structures.html#igraph-Adjlists}
}
\usage{
as_adjlist(graph, mode = 3L, loops = 1L, multiple = 1L)

as_inclist(graph, mode = 3L, loops = 1L)
}
\arguments{
\item{graph}{igraph_ptr object.}

\item{mode}{Edge type to count; \{1: OUT, 2: IN, 3: ALL\}.}

\item{loops}{How to treat loop edges; \{0: NO_LOOPS, 1: LOOPS, 2: LOOPS_ONCE\}}

\item{multiple}{Boolean, whether to keep multiple (parallel) edges.}
}
\description{
These functions are all-vertices version of \code{neighbors()} and \code{incident()}.
}
\details{
\code{as_adjlist()}: Convert a graph to a list of adjacent vertices.

\code{as_inclist()}: Convert a graph to a list of incident edges.
}
\seealso{
\code{\link[=neighbors]{neighbors()}} and \code{\link[=incident]{incident()}} for single-vertex version.
}
