% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjlist.R
\name{as_adjlist}
\alias{as_adjlist}
\alias{as_inclist}
\title{Adjacency lists}
\source{
\url{https://igraph.org/c/doc/ch07.html#igraph-Adjlists}
}
\usage{
as_adjlist(graph, mode = 3L)

as_inclist(graph, mode = 3L)
}
\arguments{
\item{graph}{IGraph object.}

\item{mode}{Edge type to count; {1: OUT, 2: IN, 3: ALL}.}
}
\description{
These functions are all-vertices version of \code{neighbors()} and \code{incident()}.
}
\details{
\code{as_adjlist()}: Convert a graph to a list of adjacent vertices.

\code{as_inclist()}: Convert a graph to a list of incident edges.
}
\seealso{
\code{\link[=neighbors]{neighbors()}} and \code{\link[=incident]{incident()}} for single-vertex version.
}
