% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{as_igraph}
\alias{as_igraph}
\alias{as_igraph.default}
\alias{as_igraph.matrix}
\alias{as_igraph.data.frame}
\alias{graph_from_data_frame}
\alias{graph_from_symbolic_edgelist}
\alias{graph_from_edgelist}
\title{Create graph from R objects}
\usage{
as_igraph(x, ...)

\method{as_igraph}{default}(x, ...)

\method{as_igraph}{matrix}(x, ...)

\method{as_igraph}{data.frame}(x, ...)

graph_from_data_frame(df, directed = TRUE)

graph_from_symbolic_edgelist(edgelist, directed = TRUE)

graph_from_edgelist(edgelist, directed = TRUE)
}
\arguments{
\item{x}{A vector, matrix, or data.frame.}

\item{...}{Additional arguments passed to the graph creation function.}

\item{df}{data.frame that includes an edgelist and edge attributes.}

\item{directed}{Boolean}

\item{edgelist}{Two-column matrix.}
}
\description{
Create graph from R objects
}
\seealso{
\code{\link[=graph_create]{graph_create()}}, \code{\link[=as.data.frame.igraph_ptr]{as.data.frame.igraph_ptr()}}
}
