% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/generators.R
\name{graph_from_data_frame}
\alias{graph_from_data_frame}
\alias{graph_from_symbolic_edgelist}
\alias{graph_from_edgelist}
\alias{as_igraph}
\alias{as_igraph.default}
\alias{as_igraph.matrix}
\alias{as_igraph.data.frame}
\title{Create graph from R objects}
\usage{
graph_from_data_frame(df, directed = TRUE)

graph_from_symbolic_edgelist(edgelist, directed = TRUE)

graph_from_edgelist(edgelist, directed = TRUE)

as_igraph(x)

\method{as_igraph}{default}(x)

\method{as_igraph}{matrix}(x)

\method{as_igraph}{data.frame}(x)
}
\arguments{
\item{df}{data.frame that includes an edgelist and edge attributes.}

\item{directed}{Boolean}

\item{edgelist}{Two-column matrix.}

\item{x}{A vector, matrix, or data.frame.}
}
\description{
Create graph from R objects
}
\seealso{
\code{\link[=graph_create]{graph_create()}}
}
