% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{as_igraph}
\alias{as_igraph}
\alias{as_igraph.default}
\alias{as_igraph.matrix}
\alias{as_igraph.data.frame}
\title{Shortcut to deterministic graph generators}
\usage{
as_igraph(x)

\method{as_igraph}{default}(x)

\method{as_igraph}{matrix}(x)

\method{as_igraph}{data.frame}(x)
}
\arguments{
\item{x}{A vector, matrix, or data.frame.}
}
\description{
Shortcut to deterministic graph generators
}
\seealso{
\code{\link[=graph_create]{graph_create()}}, \code{\link[=graph_from_edgelist]{graph_from_edgelist()}}, \code{\link[=graph_from_data_frame]{graph_from_data_frame()}}
}
