% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{vertex_attr}
\alias{vertex_attr}
\alias{vertex_attr<-}
\alias{edge_attr}
\alias{edge_attr<-}
\alias{V}
\alias{E}
\title{Get and set vertex and edge attributes}
\usage{
vertex_attr(graph, name)

vertex_attr(graph, name) <- value

edge_attr(graph, name)

edge_attr(graph, name) <- value

V(graph)

E(graph)
}
\arguments{
\item{graph}{IGraph object.}

\item{name}{Name of the attribute to query or set.
If missing, all the attributes are returned or replaced.}

\item{value}{A vector or data.frame.}
}
\description{
Get and set vertex and edge attributes
}
