% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-add-delete.R
\name{basic-add-delete}
\alias{basic-add-delete}
\alias{add_edges}
\alias{add_vertices}
\alias{delete_edges}
\alias{delete_vertices}
\title{Adding and deleting vertices and edges}
\source{
\url{https://igraph.org/c/html/latest/igraph-Basic.html#adding-and-deleting-vertices-and-edges}
}
\usage{
add_edges(graph, eids)

add_vertices(graph, n)

delete_edges(graph, eids)

delete_vertices(graph, vids)
}
\arguments{
\item{graph}{igraph_ptr object.}

\item{eids}{Edges of interest.}

\item{n}{number of vertices to add.}

\item{vids}{Vertices of interest.}
}
\value{
A modified graph.
}
\description{
Adding and deleting vertices and edges
}
