% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{basic-adjacent}
\alias{basic-adjacent}
\alias{edge}
\alias{neighbors}
\alias{incident}
\alias{degree}
\title{Basic query operations to get adjacency}
\source{
\url{https://igraph.org/c/doc/igraph-Basic.html#basic-query-operations}
}
\usage{
edge(graph, eid)

neighbors(graph, vid, mode = 3L, loops = 1L, multiple = TRUE)

incident(graph, vid, mode = 3L, loops = 1L)

degree(graph, vids = integer(0), mode = 3L, loops = 1L)
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}

\item{eid}{An integer value of edge ID.}

\item{vid}{An integer value of vertex ID.}

\item{mode}{An integer value of edge type to count; \{1: OUT, 2: IN, 3: ALL\}.}

\item{loops}{whether self-loop should be counted.}

\item{multiple}{A logical value, whether to keep multiple (parallel) edges.}

\item{vids}{An integer vector of vertex IDs.}
}
\value{
\code{edge()} returns the head and tail vertices of an edge.

\code{neighbors()} returns the adjacent vertices to a vertex.

\code{incident()} returns the incident edges of a vertex.

\code{degree()} returns the degrees of vertices in a graph.
}
\description{
Basic query operations to get adjacency
}
\examples{
g = graph_tree(5L)
edge(g, 1L)
neighbors(g, 1L)
incident(g, 1L)
degree(g, mode = 1L)
degree(g, mode = 2L)
degree(g, mode = 3L)
}
\seealso{
\code{\link[=as_adjlist]{as_adjlist()}} and \code{\link[=as_inclist]{as_inclist()}} for the all-vertices version of
\code{neighbors()} and \code{incident()}.

\code{\link[=is_sink]{is_sink()}}, \code{\link[=is_source]{is_source()}}, \code{\link[=Vsink]{Vsink()}}, \code{\link[=Vsource]{Vsource()}} for shortcuts to
get zero-degree vertices.
}
