% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\docType{data}
\name{basic-macros}
\alias{basic-macros}
\alias{IGRAPH_VCOUNT_MAX}
\alias{IGRAPH_ECOUNT_MAX}
\alias{IGRAPH_UNLIMITED}
\title{Miscellaneous macros}
\format{
An object of class \code{integer} of length 1.
}
\source{
\url{https://igraph.org/c/doc/igraph-Basic.html#misc-helper-functions}
}
\usage{
IGRAPH_VCOUNT_MAX

IGRAPH_ECOUNT_MAX

IGRAPH_UNLIMITED
}
\description{
igraphlite uses 32-bit igraph,
which limits the maximum number of vertices to \code{2 ** 31 - 2}.
}
\examples{
IGRAPH_VCOUNT_MAX

IGRAPH_ECOUNT_MAX

IGRAPH_UNLIMITED
}
\keyword{datasets}
