% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{neighbors}
\alias{neighbors}
\alias{degree}
\title{Basic interface}
\usage{
neighbors(graph, v, mode = 3L)

degree(graph, vids = numeric(0), mode = 3L, loops = TRUE)
}
\arguments{
\item{graph}{IGraph object.}

\item{v, vids}{vertices of which the degree will be calculated.}

\item{mode}{edge type to count; {1: out, 2: in, 3: all}}

\item{loops}{whether self-loop should be counted.}
}
\description{
Basic interface
}
\details{
\code{neighbors()}: Adjacent vertices to a vertex.

\code{degree()}: The degree of some vertices in a graph.
}
