% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{vcount}
\alias{vcount}
\alias{ecount}
\alias{edge}
\alias{neighbors}
\alias{incident}
\alias{is_directed}
\alias{degree}
\title{Basic Query Operations}
\source{
\url{https://igraph.org/c/doc/igraph-Basic.html}
}
\usage{
vcount(graph)

ecount(graph)

edge(graph, eid)

neighbors(graph, vid, mode = 3L)

incident(graph, vid, mode = 3L)

is_directed(graph)

degree(graph, vids = numeric(0), mode = 3L, loops = TRUE)
}
\arguments{
\item{graph}{IGraph object.}

\item{eid}{Edges of interest.}

\item{vid}{Vertices of interest.}

\item{mode}{Edge type to count; {1: OUT, 2: IN, 3: ALL}.}

\item{vids}{Vertices of interest.}

\item{loops}{whether self-loop should be counted.}
}
\description{
Basic Query Operations
}
\details{
\code{vcount()}: The number of vertices in a graph.

\code{ecount()}: The number of edges in a graph.

\code{edge()}: Gives the head and tail vertices of an edge.

\code{neighbors()}: Adjacent vertices to a vertex.

\code{incident()}: Gives the incident edges of a vertex.

\code{is_directed()}: Is this a directed graph?

\code{degree()}: The degree of some vertices in a graph.
}
\seealso{
\code{\link[=as_adjlist]{as_adjlist()}} and \code{\link[=as_inclist]{as_inclist()}} for all-vertices version.
}
