% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{basic-query}
\alias{basic-query}
\alias{vcount}
\alias{ecount}
\alias{is_directed}
\title{Basic query operations}
\source{
\url{https://igraph.org/c/doc/igraph-Basic.html#basic-query-operations}
}
\usage{
vcount(graph)

ecount(graph)

is_directed(graph)
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}
}
\value{
\code{vcount()} returns the number of vertices in a graph.

\code{ecount()} returns the number of edges in a graph.

\code{is_directed()} returns whether the graph is directed.
}
\description{
Basic query operations
}
\examples{
g = graph_tree(5L)
vcount(g)
ecount(g)
is_directed(g)
}
