% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-centrality.R
\name{centrality}
\alias{centrality}
\alias{edge_betweenness}
\alias{edge_betweenness_subset}
\title{Centrality measures}
\source{
\url{https://igraph.org/c/doc/igraph-Structural.html#centrality-measures}
}
\usage{
edge_betweenness(graph, directed = is_directed(graph), weights = numeric(0))

edge_betweenness_subset(
  graph,
  directed = is_directed(graph),
  eids = integer(0),
  weights = numeric(0),
  from = integer(0),
  to = integer(0)
)
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}

\item{directed}{A logical value, whether to consider directed paths.
Ignored for undirected graphs.}

\item{weights}{A numeric vector of edge weights; \code{TRUE} to use \code{Eattr(graph, "weight")}.}

\item{eids}{An integer vector of edge IDs.}

\item{from}{An integer vector of vertex IDs.}

\item{to}{An integer vector of vertex IDs.}
}
\value{
\code{edge_betweenness()} returns a numeric vector of edge betweenness
for each edge in the graph.

\code{edge_betweenness_subset()} is a variant of \code{edge_betweenness()}
that computes the betweenness using a subset of paths between \code{from} and \code{to}.
}
\description{
The betweenness centrality of an edge is the number of geodesics going through it.
}
\examples{
g = graph_tree(5L)
edge_betweenness(g)
edge_betweenness_subset(g, from = Vsource(g), to = Vsink(g))
}
