% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{edge_betweenness}
\alias{edge_betweenness}
\alias{edge_betweenness_subset}
\title{Centrality measures}
\source{
\url{https://igraph.org/c/doc/igraph-Structural.html#centrality-measures}
}
\usage{
edge_betweenness(graph, directed = is_directed(graph), weights = numeric(0))

edge_betweenness_subset(
  graph,
  directed = is_directed(graph),
  eids = integer(0),
  weights = numeric(0),
  from = integer(0),
  to = integer(0)
)
}
\arguments{
\item{graph}{igraph_ptr object.}

\item{directed}{Boolean, whether to consider directed paths. Ignored for undirected graphs.}

\item{weights}{Numeric edge weights; \code{TRUE} to use \code{Eattr(graph, "weight")}.}

\item{eids}{Edges of interest.}

\item{from}{Vertices of interest.}

\item{to}{Vertices of interest.}
}
\description{
Centrality measures
}
