% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-centrality.R
\name{centrality}
\alias{centrality}
\alias{betweenness}
\alias{edge_betweenness}
\alias{edge_betweenness_subset}
\title{Centrality measures}
\source{
\url{https://igraph.org/c/doc/igraph-Structural.html#centrality-measures}
}
\usage{
betweenness(
  graph,
  ...,
  weights = numeric(0),
  from = integer(0),
  to = integer(0),
  vids = integer(0),
  directed = is_directed(graph),
  normalized = FALSE,
  cutoff = -1
)

edge_betweenness(
  graph,
  ...,
  weights = numeric(0),
  from = integer(0),
  to = integer(0),
  eids = integer(0),
  directed = is_directed(graph),
  normalized = FALSE,
  cutoff = -1
)

edge_betweenness_subset(
  graph,
  ...,
  weights = numeric(0),
  from = integer(0),
  to = integer(0),
  eids = integer(0),
  directed = is_directed(graph),
  normalized = FALSE
)
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}

\item{...}{Unused, but temporarily included to avoid silent bugs and
to prompt users to cope with breaking changes in version 1.0.}

\item{weights}{A numeric vector of edge weights; \code{TRUE} to use \code{Eattr(graph, "weight")}.}

\item{from}{An integer vector of vertex IDs.}

\item{to}{An integer vector of vertex IDs.}

\item{vids}{An integer vector of vertex IDs.}

\item{directed}{A logical value, whether to consider directed paths.
Ignored for undirected graphs.}

\item{normalized}{A logical value, whether to normalize the result.}

\item{cutoff}{Maximum length of paths to be considered. Unlimited if negative.}

\item{eids}{An integer vector of edge IDs.}
}
\value{
\code{betweenness()} returns a numeric vector of betweenness
for each vertex in the graph.

\code{edge_betweenness()} returns a numeric vector of edge betweenness
for each edge in the graph.

\code{edge_betweenness_subset()} is a variant of \code{edge_betweenness()}
that computes the betweenness using a subset of paths between \code{from} and \code{to}.
}
\description{
The betweenness centrality of a vertex or an edge is the number of geodesics
going through it.
}
\examples{
g = graph_tree(5L)
betweenness(g)
edge_betweenness(g)
edge_betweenness(g, from = Vsource(g), to = Vsink(g))
edge_betweenness(g, normalized = TRUE)
edge_betweenness(g, cutoff = 1)
}
