% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{common_params}
\alias{common_params}
\title{Common parameters used throughout the package}
\usage{
common_params(
  graph,
  vids,
  vid,
  from,
  to,
  eids,
  eid,
  edges,
  mode,
  weights,
  directed,
  loops,
  multiple,
  normalized
)
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}

\item{vids}{An integer vector of vertex IDs.}

\item{vid}{An integer value of vertex ID.}

\item{from}{An integer vector of vertex IDs.}

\item{to}{An integer vector of vertex IDs.}

\item{eids}{An integer vector of edge IDs.}

\item{eid}{An integer value of edge ID.}

\item{edges}{An integer vector of vertex IDs; the first two elements are the first edge, etc.}

\item{mode}{An integer value of edge type to count; \{1: OUT, 2: IN, 3: ALL\}.}

\item{weights}{A numeric vector of edge weights; \code{TRUE} to use \code{Eattr(graph, "weight")}.}

\item{directed}{A logical value, whether to consider directed paths.
Ignored for undirected graphs.}

\item{loops}{An integer value, how to treat loop edges;
\{0: NO_LOOPS, 1: LOOPS/LOOPS_TWICE, 2: LOOPS_ONCE\}}

\item{multiple}{A logical value, whether to keep multiple (parallel) edges.}

\item{normalized}{A logical value, whether to normalize the result.}
}
\description{
This function does nothing. It is designed to be used with @inheritParams.
}
\keyword{internal}
