% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{common_params}
\alias{common_params}
\title{Common parameters used throughout the package}
\usage{
common_params(graph, vids, v, from, to, mode, weights, directed)
}
\arguments{
\item{graph}{IGraph object.}

\item{vids}{Vertices of interest.}

\item{v}{Vertices of interest.}

\item{from}{Vertices of interest.}

\item{to}{Vertices of interest.}

\item{mode}{Edge type to count; {1: OUT, 2: IN, 3: ALL}.}

\item{weights}{Numeric edge weights; \code{TRUE} to use \code{graph$Eattr$weight}.}

\item{directed}{Boolean, whether to consider directed paths. Ignored for undirected graphs.}
}
\description{
This function does nothing. It is designed to be used with @inheritParams.
}
\keyword{internal}
