% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural.R
\name{induced_subgraph}
\alias{induced_subgraph}
\title{Graph Components}
\source{
\url{https://igraph.org/c/doc/igraph-Structural.html#idm470926080896}
}
\usage{
induced_subgraph(graph, vids, impl = 0L)
}
\arguments{
\item{graph}{IGraph object.}

\item{vids}{Vertices of interest.}

\item{impl}{how to construct a new graph:
{0: AUTO, 1: COPY_AND_DELETE, 2: CREATE_FROM_SCRATCH}}
}
\description{
Graph Components
}
