% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-components.R
\name{components}
\alias{components}
\alias{subcomponent}
\alias{subcomponents}
\title{Graph Components}
\source{
\url{https://igraph.org/c/doc/igraph-Structural.html#graph-components}
}
\usage{
subcomponent(graph, vid, mode = 1L)

subcomponents(graph, vids, mode = 1L)
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}

\item{vid}{An integer value of vertex ID.}

\item{mode}{An integer value of edge type to count; \{1: OUT, 2: IN, 3: ALL\}.}

\item{vids}{An integer vector of vertex IDs.}
}
\value{
\code{subcomponent()} returns the IDs of the vertices reachable from the
given vertex.

\code{subcomponents()} returns a list of integer vectors by applying
\code{subcomponent()} for each vertex in \code{vids}.
}
\description{
Graph Components
}
\examples{
g = graph_tree(5L)
subcomponent(g, 2L)
subcomponents(g, c(1L, 2L))
}
\seealso{
\code{\link[=induced_subgraph]{induced_subgraph()}}
}
