% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural.R
\name{are_adjacent}
\alias{are_adjacent}
\title{Basic Properties}
\source{
\url{https://igraph.org/c/doc/igraph-Structural.html}
}
\usage{
are_adjacent(graph, from, to)
}
\arguments{
\item{graph}{igraph_ptr object.}

\item{from}{Vertices of interest.}

\item{to}{Vertices of interest.}
}
\description{
Basic Properties
}
