% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic-constructors.R
\name{constructors}
\alias{constructors}
\alias{graph_empty}
\alias{graph_copy}
\title{Graph constructors}
\usage{
graph_empty(n = 0L, directed = TRUE)

graph_copy(graph)
}
\arguments{
\item{n}{An integer, the number of vertices in the graph.}

\item{directed}{A logical value, whether to consider directed paths.
Ignored for undirected graphs.}

\item{graph}{An \code{\link{igraph_ptr}} object.}
}
\value{
An \code{\link{igraph_ptr}} object.
}
\description{
\code{graph_empty()} creates an empty graph with a specified number of vertices.
An empty graph in \emph{igraph} is a graph with no edges, which may have vertices.

\code{graph_copy()} creates a deep copy of the graph.
Normal assignment like \code{g2 = g1} creates a shallow copy, which means that
the internal data is shared between the two R objects, and the modification
on one of them will affect the other.
}
\examples{
g = graph_empty(4L)
g

g_shallow = g
g_deep = graph_copy(g)
add_edges(g, seq_len(4L))

g_shallow

g_deep
}
