% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.R
\name{degree}
\alias{degree}
\title{The degree of some vertices in a graph}
\usage{
degree(graph, vs = NA, mode = c("all", "out", "in", "total"),
  loops = TRUE)
}
\arguments{
\item{graph}{IGraph object.}

\item{vs}{vertices of which the degree will be calculated.}

\item{mode}{type of the degree.}

\item{loops}{whether self-loop should be counted.}
}
\description{
This function calculates the in-, out- or total degree of the specified vertices.
}
