% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{find_common_ancestors}
\alias{find_common_ancestors}
\alias{upstream_vertices}
\title{Experimental functions}
\usage{
find_common_ancestors(graph, vids)

upstream_vertices(graph, vids, to_mrca = TRUE)
}
\arguments{
\item{graph}{igraph_ptr object.}

\item{vids}{Vertices of interest.}

\item{to_mrca}{Boolean.}
}
\description{
Experimental functions
}
