% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{graph_from_data_frame}
\alias{graph_from_data_frame}
\alias{graph_from_symbolic_edgelist}
\alias{graph_from_edgelist}
\alias{graph_create}
\alias{graph_tree}
\title{Deterministic Graph Generators}
\source{
\url{https://igraph.org/c/doc/igraph-Generators.html}
}
\usage{
graph_from_data_frame(df, directed = TRUE)

graph_from_symbolic_edgelist(edgelist, directed = TRUE)

graph_from_edgelist(edgelist, directed = TRUE)

graph_create(edges, n = 0L, directed = TRUE)

graph_tree(n, children = 2L, mode = 0L)
}
\arguments{
\item{df}{data.frame that includes an edgelist and edge attributes.}

\item{directed}{Boolean}

\item{edgelist}{Two-column matrix.}

\item{edges}{The edges to add, the first two elements are the first edge, etc.}

\item{n}{Integer, the number of vertices in the graph.}

\item{children}{Integer, the number of children of a vertex in the tree.}

\item{mode}{\code{igraph_tree_mode_t} {1: OUT, 2: IN, 3: UNDIRECTED}}
}
\description{
Deterministic Graph Generators
}
