% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraphlite-package.R
\docType{class}
\name{igraph_ptr}
\alias{igraph_ptr}
\title{igraph_ptr class: thin wrapper of \code{igraph_t}.}
\format{
external pointer to C++ IGraph class object.
}
\source{
\url{https://github.com/igraph/igraph/blob/master/include/igraph_datatype.h}
}
\description{
igraph_ptr class: thin wrapper of \code{igraph_t}.
}
\seealso{
\code{\link[=graph_from_edgelist]{graph_from_edgelist()}}, \code{\link[=graph_from_data_frame]{graph_from_data_frame()}}, \code{\link[=graph_tree]{graph_tree()}}
for object instantiation.
}
\keyword{data}
