% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraphlite-package.R
\docType{class}
\name{igraph_ptr}
\alias{igraph_ptr}
\title{igraph_ptr class: thin wrapper of \code{igraph_t}.}
\format{
external pointer to C++ IGraph class object.
}
\source{
\url{https://github.com/igraph/igraph/blob/master/include/igraph_datatype.h}
}
\description{
igraph_ptr class: thin wrapper of \code{igraph_t}.
}
\seealso{
\code{\link[=is_igraph]{is_igraph()}} to check if an object is of class \code{igraph_ptr}.

\code{\link[=as_igraph]{as_igraph()}} and \code{\link[=as.data.frame.igraph_ptr]{as.data.frame.igraph_ptr()}} for conversion
from and to R objects.

\code{\link[=graph_lattice]{graph_lattice()}} and \code{\link[=graph_tree]{graph_tree()}} for deterministic graph generation.
}
\keyword{data}
