% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraphlite-package.R
\docType{package}
\name{igraphlite-package}
\alias{igraphlite}
\alias{igraphlite-package}
\title{igraphlite: Lightweight R Interface to Igraph Network Analysis Library}
\description{
This is an Rcpp-powered R package interface to igraph network analysis
    library. It is not a fork of the existing R igraph package, but a reboot
    from scratch striving for good performance, simple codebase, and the
    consistency with the C implementation.
}
\author{
\strong{Maintainer}: Watal M. Iwasaki \email{heavy.watal@gmail.com}

}
\keyword{internal}
