% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraphlite-package.R
\docType{package}
\name{igraphlite-package}
\alias{igraphlite}
\alias{igraphlite-package}
\title{igraphlite: Lightweight R Interface to Igraph Network Analysis Library}
\description{
This is an Rcpp-powered R package interface to igraph network analysis
    library. It is not a fork of the existing R igraph package, but a reboot
    from scratch striving for good performance, simple codebase, and the
    consistency with the C implementation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://heavywatal.github.io/rigraphlite/}
  \item \url{https://github.com/heavywatal/rigraphlite}
  \item Report bugs at \url{https://github.com/heavywatal/rigraphlite/issues}
}

}
\author{
\strong{Maintainer}: Watal M. Iwasaki \email{heavywatal@gmail.com} (\href{https://orcid.org/0000-0002-4718-7605}{ORCID})

}
\keyword{internal}
