% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraphlite-package.R
\name{is_igraph}
\alias{is_igraph}
\title{Check if an object is igraph_ptr class}
\usage{
is_igraph(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
\code{TRUE} if the object is of class \code{igraph_ptr}, \code{FALSE} otherwise.
}
\description{
Check if an object is igraph_ptr class
}
