% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_nicely}
\alias{layout_nicely}
\alias{layout_random}
\alias{layout_drl}
\alias{layout_fruchterman_reingold}
\alias{layout_mds}
\alias{layout_reingold_tilford}
\alias{layout_reingold_tilford_circular}
\title{Layouts for Graph Drawing}
\source{
\url{https://igraph.org/c/doc/igraph-Layout.html}
}
\usage{
layout_nicely(graph, ...)

layout_random(graph)

layout_drl(graph)

layout_fruchterman_reingold(graph, grid = 2L)

layout_mds(graph)

layout_reingold_tilford(graph, mode = 1L, roots = integer(0))

layout_reingold_tilford_circular(graph, mode = 1L, roots = integer(0))
}
\arguments{
\item{graph}{igraph_ptr object.}

\item{...}{Passed to an internal layout function.}

\item{grid}{Whether to use the (fast but less accurate) grid based version;
\{0: GRID, 1: NOGRID, 2: AUTOGRID\}.}

\item{mode}{Edge type to count; \{1: OUT, 2: IN, 3: ALL\}.}

\item{roots}{Root vertices.}
}
\description{
Layouts for Graph Drawing
}
