% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural.R
\name{neighborhood_size}
\alias{neighborhood_size}
\alias{neighborhood}
\title{Neighborhood of vertices}
\source{
\url{https://igraph.org/c/doc/igraph-Structural.html}
}
\usage{
neighborhood_size(graph, vids = numeric(0), order = 1L, mode = 1L,
  mindist = 0L)

neighborhood(graph, vids = numeric(0), order = 1L, mode = 1L,
  mindist = 0L)
}
\arguments{
\item{graph}{IGraph object.}

\item{vids}{Vertices of interest.}

\item{order}{integer}

\item{mode}{Edge type to count; {1: OUT, 2: IN, 3: ALL}.}

\item{mindist}{The minimum distance to include a vertex in the counting.}
}
\description{
Neighborhood of vertices
}
