% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural-paths.R
\name{paths}
\alias{paths}
\alias{get_shortest_paths}
\alias{get_all_shortest_paths}
\alias{get_all_simple_paths}
\title{Shortest Path Related Functions}
\source{
\url{https://igraph.org/c/doc/igraph-Structural.html#shortest-path-related-functions}
}
\usage{
get_shortest_paths(
  graph,
  ...,
  weights = numeric(0L),
  from = integer(0L),
  to = integer(0L),
  mode = 3L
)

get_all_shortest_paths(
  graph,
  ...,
  weights = numeric(0L),
  from = integer(0L),
  to = integer(0L),
  mode = 3L
)

get_all_simple_paths(
  graph,
  from,
  to = integer(0L),
  mode = 3L,
  minlen = -1L,
  maxlen = -1L,
  max_results = -1L
)
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}

\item{...}{Unused, but temporarily included to avoid silent bugs and
to prompt users to cope with breaking changes in version 1.0.}

\item{weights}{A numeric vector of edge weights; \code{TRUE} to use \code{Eattr(graph, "weight")}.}

\item{from}{An integer vector of vertex IDs.}

\item{to}{An integer vector of vertex IDs.}

\item{mode}{An integer value of edge type to count; \{1: OUT, 2: IN, 3: ALL\}.}

\item{minlen}{Minimum length of paths to be considered. Unlimited if negative.}

\item{maxlen}{Maximum length of paths to be considered. Unlimited if negative.}

\item{max_results}{At most this many paths will be recorded. Unlimited if negative.}
}
\value{
A list of integer IDs of vertices in the paths.
}
\description{
\code{get_shortest_paths()} returns a single shortest path per vertex pair,
while the variant with \verb{_all_} returns all found paths.
The variant with \verb{_simple_} removes paths with cycles.
The results should be the same for simple graphs like trees, but the order of
the paths may differ.
}
\examples{
g = graph_tree(5L)
vsource = Vsource(g)
vsink = Vsink(g)
get_shortest_paths(g, from = vsource, to = vsink, mode = 1L)
get_all_shortest_paths(g, from = vsource, to = vsink, mode = 1L)
get_all_simple_paths(g, from = vsource, to = vsink, mode = 1L)
}
