% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-method.R
\name{augment.igraph_ptr}
\alias{augment.igraph_ptr}
\alias{plot.igraph_ptr}
\title{Methods for quick visualization}
\usage{
\method{augment}{igraph_ptr}(x, layout = NULL, ...)

\method{plot}{igraph_ptr}(x, ..., lwd = 0.6, cex = 6, col = "#cccccc", pch = 16)
}
\arguments{
\item{x}{An igraph_ptr object.}

\item{layout}{A function or resulting data.frame.
If not provided, \code{\link[=layout_nicely]{layout_nicely()}} is applied.}

\item{...}{Extra arguments passed to the layout function.}

\item{lwd}{Passed to \code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}}.}

\item{cex, col, pch}{Passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}} and \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}.}
}
\value{
\code{augment()} returns a data frame to be used with \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.

\code{plot.igraph_ptr()} returns a ggplot object.
}
\description{
Methods for quick visualization
}
\examples{
g = graph_tree(5L)
augment(g, layout = layout_reingold_tilford)

plot(g, layout = layout_reingold_tilford) +
  ggplot2::theme_minimal(base_size = 14)
}
