% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-method.R
\name{augment.Rcpp_IGraph}
\alias{augment.Rcpp_IGraph}
\alias{plot.Rcpp_IGraph}
\title{Methods for quick visualization}
\usage{
\method{augment}{Rcpp_IGraph}(x, layout = NULL, ...)

\method{plot}{Rcpp_IGraph}(x, ..., lwd = 0.5, cex = 5, col = "#cccccc", pch = 16)
}
\arguments{
\item{x}{Rcpp_Igraph object.}

\item{layout}{A function or resulting data.frame.
If not provided, \link{layout_nicely} is applied.}

\item{...}{passed to \link{augment} or layout function.}

\item{lwd}{passed to \link[ggplot2:geom_segment]{ggplot2::geom_segment}.}

\item{cex, col, pch}{passed to \link[ggplot2:geom_point]{ggplot2::geom_point} and \link[ggplot2:geom_text]{ggplot2::geom_text}.}
}
\description{
Methods for quick visualization
}
