% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraphlite-package.R
\name{igraph_rng_seed}
\alias{igraph_rng_seed}
\title{Seeds a random number generator}
\source{
\url{https://igraph.org/c/doc/igraph-Random.html}
}
\usage{
igraph_rng_seed(seed)
}
\arguments{
\item{seed}{An integer.}
}
\value{
\code{NULL}.
}
\description{
Initializes \code{igraph_rng_default()} with an integer seed.
}
\examples{
igraph_rng_seed(24601L)
}
