% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rng_seed}
\alias{rng_seed}
\title{Random numbers}
\source{
\url{https://igraph.org/c/doc/igraph-Random.html}
}
\usage{
rng_seed(seed)
}
\arguments{
\item{seed}{An integer.}
}
\description{
Random numbers
}
