% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-method.R
\name{print.igraph_ptr}
\alias{print.igraph_ptr}
\title{S3 methods to get and view internal data.}
\usage{
\method{print}{igraph_ptr}(x, ...)
}
\arguments{
\item{x}{An igraph_ptr object.}

\item{...}{Extra arguments passed to the method.}
}
\value{
The original \code{x} invisibly.
}
\description{
S3 methods to get and view internal data.
}
\examples{
print(graph_tree(5L))
}
