% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-method.R
\name{print.igraph_ptr}
\alias{print.igraph_ptr}
\title{S3 methods to get and view internal data.}
\usage{
\method{print}{igraph_ptr}(x, ...)
}
\arguments{
\item{x}{igraph_ptr object.}

\item{...}{passed to methods.}
}
\description{
S3 methods to get and view internal data.
}
