% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural.R
\name{distances}
\alias{distances}
\alias{get_shortest_paths}
\alias{get_all_shortest_paths}
\alias{get_all_simple_paths}
\alias{average_path_length}
\alias{path_length_hist}
\title{Shortest Path Related Functions}
\source{
\url{https://igraph.org/c/doc/igraph-Structural.html}
}
\usage{
distances(
  graph,
  from = integer(0L),
  to = from,
  weights = numeric(0L),
  mode = 3L,
  algorithm = c("dijkstra", "bellman-ford", "johnson")
)

get_shortest_paths(
  graph,
  from,
  to = integer(0L),
  weights = numeric(0L),
  mode = 3L
)

get_all_shortest_paths(
  graph,
  from,
  to = integer(0L),
  weights = numeric(0L),
  mode = 3L
)

get_all_simple_paths(graph, from, to = integer(0L), cutoff = -1L, mode = 3L)

average_path_length(
  graph,
  weights = numeric(0L),
  directed = FALSE,
  unconn = TRUE
)

path_length_hist(graph, directed = FALSE)
}
\arguments{
\item{graph}{igraph_ptr object.}

\item{from}{Vertices of interest.}

\item{to}{Vertices of interest.}

\item{weights}{Numeric edge weights; \code{TRUE} to use \code{Eattr(graph, "weight")}.}

\item{mode}{Edge type to count; \{1: OUT, 2: IN, 3: ALL\}.}

\item{algorithm}{character}

\item{cutoff}{integer}

\item{directed}{Boolean, whether to consider directed paths. Ignored for undirected graphs.}

\item{unconn}{logical}
}
\description{
Shortest Path Related Functions
}
