% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink.R
\name{is_sink}
\alias{is_sink}
\alias{is_source}
\alias{Vsink}
\alias{Vsource}
\title{Shortcuts to get sink/source vertices.}
\usage{
is_sink(graph)

is_source(graph)

Vsink(graph)

Vsource(graph)
}
\arguments{
\item{graph}{igraph_ptr object.}
}
\value{
\code{is_sink()} and \code{is_source()} return a logical vector indicating
whether the vertices in the graph are terminating or starting points,
respectively.

\code{Vsink()} and \code{Vsource()} return the terminating and starting
vertices of a graph, respectively.
a graph, respectively.
}
\description{
Shortcuts to get sink/source vertices.
}
