% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural.R
\name{shortest_paths}
\alias{shortest_paths}
\alias{neighborhood_size}
\alias{neighborhood}
\alias{induced_subgraph}
\title{Structural Properties of Graphs}
\usage{
shortest_paths(graph, from = numeric(0L), to = from,
  weights = numeric(0L), mode = 3L, algorithm = c("auto",
  "unweighted", "dijkstra", "bellman-ford", "johnson"))

neighborhood_size(graph, vids = numeric(0), order = 1L, mode = 1L,
  mindist = 0L)

neighborhood(graph, vids = numeric(0), order = 1L, mode = 1L,
  mindist = 0L)

induced_subgraph(graph, vids, impl = 0L)
}
\arguments{
\item{graph}{IGraph object.}

\item{weights}{numeric edge weights; TRUE to use graph$E$weight}

\item{mode}{edge type to count; {1: OUT, 2: IN, 3: ALL}}

\item{algorithm}{character}

\item{vids, from, to}{vertices of interest}

\item{order}{integer}

\item{mindist}{The minimum distance to include a vertex in the counting.}

\item{impl}{how to construct a new graph:
{0: AUTO, 1: COPY_AND_DELETE, 2: CREATE_FROM_SCRATCH}}
}
\description{
Structural Properties of Graphs
}
