% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraphlite-package.R
\name{igraph_version}
\alias{igraph_version}
\title{Check and return igraph version.h}
\source{
\url{https://igraph.org/c/doc/igraph-Nongraph.html}
}
\usage{
igraph_version()
}
\description{
Check and return igraph version.h
}
