% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{Vnames}
\alias{Vnames}
\alias{as_vids}
\alias{as_vnames}
\title{Conversion between vertex IDs and names}
\usage{
Vnames(graph)

as_vids(graph, vnames)

as_vnames(graph, vids)
}
\arguments{
\item{graph}{An \code{\link{igraph_ptr}} object.}

\item{vnames}{Vertex names stored in \code{Vattr(g)$name}.}

\item{vids}{An integer vector of vertex IDs.}
}
\value{
\code{Vnames()} and \code{as_vnames()} return a vector of vertex names or IDs.

\code{as_vids()} returns an integer vector of vertex IDs.
}
\description{
\code{Vnames()} is a faster alternative to \code{\link[=Vattr]{Vattr()}} to get the "name" attribute,
falling back to \code{V()} when it is not set.
\code{as_vids()} and \code{as_vnames()} provide convenient shorthands using \code{Vnames()}.
}
\examples{
g = graph_create(letters[1:6])
Vnames(g)
as_vids(g, c("d", "b"))
as_vnames(g, c(4L, 2L))

# Falls back to V(), not NULL
Vnames(graph_empty(3L))
}
